Added authorization headers to frontend requests

+++ frontend/src/components/BookDetails.js
@@ -1,13 +1,21 @@
 import React from 'react';
 import { useBookContext } from '../hooks/useBookContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 export default function BookDetails({ book }) {
   const { dispatch } = useBookContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
     const response = await fetch('http://localhost:4000/api/books/' + book._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json();
